Probe scan: Improve code organization + first cleanup pass
