+++ carveracontroller/addons/probe_scan/core/__init__.py
@@ -1,0 +1,1 @@
+"""Probe scan core: session model, geometry, G-code, feature resolution, I/O."""
