+++ carveracontroller/addons/probe_scan/ui/__init__.py
@@ -1,0 +1,1 @@
+"""Probe scan UI: popup, sketch, display formatting, probe lifecycle."""
