+++ carveracontroller/makera.kv
@@ -4,7 +4,7 @@
 #:import tr carveracontroller.__main__.tr
 #:include addons/tooltips/Tooltips.kv
 #:include addons/probing/ProbingPopup.kv
-#:include addons/probe_scan/ProbeScanPopup.kv
+#:include addons/probe_scan/ui/ProbeScanPopup.kv
 
 <Label>:
     font_name: 'ARIALUNI'
